Edit FavoritesPresenter to parse favorites list array
FavoritesPresenter now takes an array argument and loops through each band object in the aray. For each object, band container, name, image, and remove nodes are created and appended to favoritesMenuNode.

FavoritesPresenter previously took a band object as an argument. This resulted in appending only one band object to the favorites list and the favorites list only existing as a div in the dom, not an object.