Add favorites create and index methods
When user visits site, guest user is created with users's geolocation.
User can add favorite artists, which will be saved in the database.
When the user enters his/her phone number, the account will become
a registered user. If conversion does not happen, the account will
be deleted in a nightly rake task.