Add logout functionality
if user is registered he/she will see a logout button on favorites
menu instead of a notify button. Clicking notify button will
destroy the user session.

Notify is uncessary if user is registered because favorite artists
are saved in the database as they are entered

Add css style for logout button inherit from notify button

Need to implement a flash notice that informs user of this autosave